+++ src/main/java/amp/awec/WorldEditMod.java
@@ -79,6 +79,7 @@ public class WorldEditMod implements ModInitializer {
 		CommandManager.registerCommand(new CommandGmask());
 		CommandManager.registerCommand(new CommandDistr());
 		CommandManager.registerCommand(new CommandCount());
+		CommandManager.registerCommand(new CommandTrim());
 
 
 		if (FabricLoader.getInstance().getEnvironmentType() == EnvType.SERVER) {